requirements.txt
py-trello
python-slugify
urllib3
boto3
spacy
